+++ apps/autotest/scripts/rootModuleTest.js
@@ -1,0 +1,3 @@
+module.exports.root = function () {
+  console.log('require("/scripts/rootModuleTest") works, so NODE_PATh is set to configPath')
+}